First upload
I created a little controller with 2 potentiometers to control angle and velocity by hand. The frequency of reading and publishing on serial is about 2ms, but can be put to 1ms. Also created a WindowsForm, I choose this strategy only because I already used it to test serial com. The forms converts the two potentiometers data into valuable information for the rover and publish everything on mqtt.